app/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
google-cloud-kms>=2.21,<3.0
google-cloud-firestore>=2.16,<3.0
google-cloud-storage>=2.14,<3.0
google-cloud-logging>=3.9,<4.0
firebase-admin>=6.4,<7.0
pydantic>=2.10,<3.0
cryptography>=44.0,<45.0
llama-cpp-python==0.3.4
scipy>=1.14,<2.0
httpx>=0.27,<1.0
websockets>=13.0,<14.0
